DB_Challenge/requirements.txt
streamlit
google-cloud-vision
Pillow
requests
sqlite3
datetime
Image
